update map component
